don't include bt_handler or killpg in backtraces
Summary: These are pretty much in every backtrace and @markw65 and I spent ~15 mins staring at why it could segfault in the `bt_handler`. It is just a symptom of where we grab the backtrace.

I'm not sure if there is a better way to do arrays without using the heap. Anyone better at C++ want to help out?

Reviewed By: @markw65

Differential Revision: D1375137