EZC: After ArrayData::copy(), reproxy the array
ProxyArray::Copy() causes the array to be unproxied, which I think is
correct, since we don't really want EZC hacks to leak out of EZC any
further than is necessary. But this does mean that callers of copy()
internal to EZC need to reproxy the array. This was already done in
zBoxAndProxy(), but was missed in ZArrVal and _zval_copy_ctor_func.