Exit server on hhconfig change
Summary: The hhi embedding option is handled at server init time, and
it's going to be more trouble than it's worth IMO to allow runtime
twiddling of it. (Off -> on could probably be made to work, but on ->
off is trickier with current machinery.) So instead, just kill the
server when we detect an hhconfig change; `hh_client` will start it up
again next time.

Reviewed By: @elgenie

Differential Revision: D1376096