Refactor JS utils
Summary: The JS version of Hack cannot make system calls (understandably).
The way we dealt with this so far was by having 2 versions of utils.ml.
The JS version was stripping out the primitives making system calls.

There was 2 problems:
1) js/utils.ml was not only stripping out the primitives making system calls,
it was stubbing modules that behave differently in JS (Path, Hhi).
2) We constantly have to modify both version of utils

This diff introduces a new module sys_utils.ml that is included by utils.ml
and regroups all the primitives making system calls.

To make this work, I had to rename all the uses of Json to Hh_json.
We had this weird situation where Json meant one thing in native world and
another in the JS world (because Json is a builtin module in the JS world).
So Hh_json was our version of Json for JS and non-existant for the native code.

I changed all of that so everybody uses Hh_json.

Reviewed By: @jwatzman

Differential Revision: D1377582