Adapt typing_suggest to new error system
Summary: The new error system doesn't throw when an error is encountered.
This breaks the suggestion mode logic.
All I did is make the code throw when an error is found, instead of
raising Error, we raise Exit (a predefined OCaml exception).
The output is slightly different (No errors! is printed in a different place),
so I fixed the .exp in the unit tests

Reviewed By: @jwatzman

Differential Revision: D1377683