Refactor constant lookup JIT helpers
There are two helpers for looking up constants in the JIT.
`lookupCnsHelper`, for getting constants without a fallback, and
`lookupCnsUHelper` for constants with a fallback. Each function did their jobs
slightly differently, with lookupCnsUHelper actually having a bug in the
fallback case.

This commit refactors the two helpers into a single `lookupCnsCommonHelper`
helper. Which should help keep things in sync.

Fixes #2544