Speed up refcount-opts
Summary: We were spending a lot of time copying the State struct, which had 4
smart::hash_maps in it. I converted two of the three sites to move instead of
copy and switched to smart::flat_map. flat_map is significantly faster to copy
and move, and isn't much slower to insert/look up at the relatively small sizes
these maps tend to be. The net effect is that optimizeRefcounts() is down to
~31% of time in optimize(), from ~35%. It's also no longer the most expensive
optimizatin - that title now belongs to reoptimize().

collectLdStacks has been dead since we removed the simplifier optimization that
replaced some of SpillStack/Call's sources with None so I removed it.

Reviewed By: @jdelong

Differential Revision: D1378844