Reorganize func.h
Summary: Reorganize func.h by "grouping similar functionalities" and "separating
interface from implementation".  Also add "documentation" to help people
decipher obscure naming conventions and PHP semantics.

Functions that are not called by anyone are removed.  I've yet to remove
the m_id property of Func::SharedData, although it appears not to
actually be used by anyone, at least through the Func::id() method.  (I
removed the method, though.)

Also turn ParamInfo into a public struct rather than a private one which
does absolutely nothing except implement getters and setters.  If anyone
feels strongly about not doing this, I can just revert the changes to
the non-func.* files, but these info structs are pretty much just meant
to be bags of data, so I don't see what the methods buy us but clutter.

This reorganization and refactor is not done (see the "Junk" section at
the bottom of Func, and the FuncEmitter which ought to be separated
out), but it's ~most of the work and I want to get it in as fast as I
can before I get merge conflicts.

Reviewed By: @jdelong

Differential Revision: D1379054