Typecheck parameters in PHP constructor invocations
Summary: NOTE: I assume there's something wrong with this fix since it was basically
the fix from D750432 which was reverted in
rHACK49fd0c13388ec40fe57b5793e301ccb61ba3b742 - but I'm not sure why.
Raising as an RFC; this bug bit me in www code today so there's
something to fix and probably a better way :-) .

Parameters weren't always checked in constructor invocations.

Specifically: calling ##new DerivedClass(...)##, where ##DerivedClass##
is a Hack class that extends a PHP class but inherits the constructor
from the PHP class. In that case the arguments are never visited.

Reviewed By: @pikatchu

Differential Revision: D1374187