Fix bug in preOptimizeAssertTypeOp, tighten up typeMightRelax logic
Summary: preOptimizeAssertTypeOp has some logic to avoiding inserting redundant
AssertTypeOps when the source value comes from an AssertTypeOp with an
appropriate typeParam. It was looking at the dest type instead of the typeParam
of the new Assert op and missing out on lots of cases.

typeMightRelax would previously only return false if guard relaxation was off
or if the value was a constant. I've extended it to do the right thing for each
dest type in the opcode table. This prevents us from doing silly things like
keeping around "AssertType<Cell> t4:Str" when t4 comes from a Concat operation.

Together, these two fixes drastically reduce the number of AsssertTypes we end
up emitting into the initial translation. This eliminates some O(n^2) guard
relaxation crap we were running into in the new test. It looks like there's
still something superlinear going on in a different part of the jit so I'll dig
into that next.

Reviewed By: @bertmaher

Differential Revision: D1380326