Fix assert in cse
Summary: The simplifier removes Movs as instruction sources as
it goes along, and as it processes later instructions, it asserts
that this has been done to all previous instructions. But if the
result of simplification is a Mov, and the dst is different from
the input, it would generate a Mov, causing the assert to fire.

Reviewed By: @bertmaher

Differential Revision: D1380875