Remove some race conditions in Class destruction
Summary: When a Unit is unloaded, it marks all its classes as
'zombies', but can't free the memory yet because there may still be
subclasses of that Class that aren't freed yet.  We had code here that
would try to release references to other Classes that a Class depends
on as early as possible (we wait for the Treadmill to do the actual
atomic release to make sure the class pointer stays available for any
currently-running requests).  As far as I can tell the only part of
this code that was required for correctness is the part that nulls out
Func*'s in the method table that the Class doesn't own (so the
destructor will not free them).

On the other hand, this code isn't thread safe: the Class loading code
chases through the linked list of defined classes trying to find a
match, and calls avail() on each one.  This function doesn't have any
synchronization around the accesses to m_declInterfaces, m_usedTraits,
or m_parent.get(), which may have been concurrently written to by
another thread running destroy().  It's also unclear to me why it is
safe to change the method table (with or without proper
synchronization)---do we we know no other thread could be loading one
of the Func*'s out of the table concurrently and trying to put it into
an ActRec?

Change this to do the releaseRefs call at treadmill time, when we know
no other threads could be reading the Class, except to check
isZombie() when it is the parent of some other class.

Reviewed By: @markw65

Differential Revision: D1381570