Converting ArrayObjcets to an array was not working correctly when the STD_PROP_LIST flag was set
Summary: The STD_PROP_LIST flag for array objects should treat properties of the array object as if they are the storage of the class. When we were converting to an array the internal storage array was still being copied. This also a problem with json_encoding where the properties were not being encoded. Fixes #2139

Reviewed By: @ptarjan

Differential Revision: D1382324