FastCGI had a bug where as if the FCGI_STDIN packet was split into two
Summary: records it would only accept the first one therefore truncating the body.

I'm not sure if this is the correct thing to do but as the string was
getting read from the chain already I suspect this is what was meant
to happen to begin with.
Closes #2933

Reviewed By: @ptarjan

Differential Revision: D1382263

Pulled By: @JoelMarcey