Plausibly fix a race condition in the PhpFile unloading code path
Summary: After a PhpFile's reference count goes to zero, we put it on
the Treadmill, but it's still in the s_md5Files map.  Another thread
might load a file with the same md5 (in computeMd5), and then lead to
incRefing the PhpFile and using it again, and we'll hit the
always_assert in ~PhpFile when the treadmill runs.  Instead, try to
get it out of the md5files map earlier---this isn't quite enough
because another thread could still pull it out, so lookup sites also
need to check if the PhpFile has a non-zero reference count when it
comes out of the s_md5Files map.

Reviewed By: @mxw

Differential Revision: D1383518