Make lookupPhpFile and lookupIncludeRoot traffic in Unit* instead of PhpFile*
Summary: All anyone does with PhpFiles these days is get the Unit* out
of it.  (They used to reference count them, but Mark fixed that a
while back, and now it's a bit redundant.)  Trying to get them out of
the m_evaledFilesMap requires more testing (to ensure the keys are the
same as phpFile->getFileName)---maybe later.

Reviewed By: @mxw

Differential Revision: D1383936