Fix memory leak in async function exception handling
Summary: If the exception was thrown in a resumed async function, the
AsyncFunctionWaitHandle::fail() incremented the reference count,
but the unwinder did not decrement it.

Let's avoid unnecessary pair of incref()/decref(), document that fail()
is supposed to consume reference and split fail() into fail() +
failCpp().

Also, avoid running PHP code when C++ exception is being handled.

Reviewed By: @jdelong

Differential Revision: D1384176