Fix occasional segfault during C++ exception unwinding
Summary: If the request times out while resumed async function is being executed,
the async function may reach Await opcode, which marks the async
function as blocked before executing resume hook. The resume hook then
throws C++ exception and unwinder marks the async function as failed.

The problem is that the async function was already marked as blocked
and its child will eventually try to unblock it, which is not a valid
operation for already finishes wait handle. This causes a segfault.

Let's fix it by simply not marking the async function as failed if it
already transitioned to another state.

Reviewed By: @jdelong

Differential Revision: D1384240