Fix a bug in how optimizeActRecs treats ReDefSP, nested inline calls broke
Summary: The code for adjusting a ReDefSP that depends on an
eliminatable DefInlineFP was written only considering the case where
it is the first ReDefSP (i.e. the one that comes at the start of the
inlined callee, not the one after an InlineReturn).  After an
InlineReturn, we could still have a ReDefSP that defines a stack
relative to a DefInlineFP if we had a nested inlined call---but the
logic for adjusting stack offsets was not correct, and it is also
incorrect to change the ReDefSP's stack source, because some user of
the ReDefSP may need values from that stack.  (The point of changing
it was because in the first ReDefSP it points to the SpillFrame for
the inlined call, and we know we can eliminiate that if the frame for
it is going away too---this can't trivially be done for nested calls
though, because there might be other ReDefSPs that depend on the stack
offset including the SpillFrame.)

Reviewed By: @swtaarrs

Differential Revision: D1384388