Break more dependencies on PhpFile* outside of file-repository.cpp
Summary: EvaledFilesMap and debugger stuff still used these, but only
to call getFileName().  After some investigation, it looks like
phpFile->getFileName() is always equivalent to
phpFile->unit()->filepath() (we set the unit's filepath to the same
string we're going to pass the PhpFile constructor), even with
Eval.CheckSymLink on, so we can skip it.  While looking through this I
noticed m_evaledFilesOrder was being added to even if we're doing a
non-initial include of a file we've already included via a different
path, which makes the results of get_included_files not match php5, so
this diff also fixes that.

Reviewed By: @alexmalyshev

Differential Revision: D1386883