Introduce kIntMapKind for IntMap array with only integer keys
Summary: This diff introduces a new ArrayKind kIntMapKind which has integer keys and is able to gracefully downgrade into a MixedArray. It templatizes the existing functions on MixedArray where necessary to raise warnings, and switches on the template parameter so that there should be no overhead for the kMixedKind case besides the extra code size.

Reviewed By: @paroski

Differential Revision: D1386055