Implemented xdebug configuration options (take 2)
Summary: This is just D1378660 fixed so that the enable test doesn't run in
RepoAuthoritative mode. The enable test fails in RepoAuthoritative mode, but
unit tests didn't run in D1378660, so this was missed. I'll make sure unit
tests run this time.

This adds the skeleton required for all the defined xdebug configuration options.
All options are bound to and set to the default values defined at
http://xdebug.org/docs/all_settings.

Since xdebug by default overloads some functions, I added a custom setting
"hhvm.xdebug-not-done.enable", that only registers xdebug functions when
set to true. By default it's false.

Since there are so many different options, I made heavy use of macros to
avoid repetitive code- I hope the style is acceptable?

I also moved some of the xdebug definitions into a header, to reduce the
size of the cpp file a bit.

Reviewed By: @ptarjan

Differential Revision: D1380781