Rewrite non-repo version of the thing formerly known as FileRepository
Summary: It was buggy, and had evolved poorly.

The main bug related to the ~PhpFile codepaths and the md5 map.  A treadmilled
PhpFile could be reused even though its unit was scheduled to be deleted.  This
diff fixes the race by having clearer ownership, and eliminating the md5 map.

Rationale on not reimplementing the md5-based cache for now: we are including
the file's path into the md5 to avoid sharing the "same" unit for files with
different paths (it would break the "Dir" and "File" opcodes as currently
implemented).  We also already have a cache to avoid starting from scratch if
the contents are the same (the Repo), so this only only can save on duplicating
JIT time/space, and Unit load time.  It also shouldn't be hitting very often: I
think the main case is if you manage to have the mtime change on a file without
the file contents actually changing.  Multiple copies of the "same" file for
different www checkouts on a shared server won't be getting shared Units right
now, and switching branches and then switching back can only reuse the Unit if
it either goes through the buggy code path and reused it even though it was
treadmilled or if the unit wasn't loaded since the time you were on the
previous branch.  So, I'm thinking let's make it simpler now and optimize it
again if we need to later.

Reviewed By: @swtaarrs

Differential Revision: D1386893