Rename file-repository to unit-cache
Summary: This is just a rename diff.  It's on top of everything else so we can
choose different names without me needing to rebase, if anyone has a better
naming idea.  This seems like a reasonable name, although both resolveVmInclude
and mangleUnitMd5 feel slightly out of place now (but resolveVmInclude (and
resolve_include) are pretty convoluted for historical reasons and deserving of
their own cleanup pass anyway, which I'm going to consider out of scope for now
since it's not causing bugs as far as we know ...).

Reviewed By: @mxw

Differential Revision: D1386894