CallsNative flag on ir opcodes should imply MayRaiseError by default
Summary: We've been burned in the past by opcodes that can throw but weren't
marked as such. MayRaiseError is now opt-out on CallsNative opcodes, by using
NNE instead of N. I tried to audit each one that broke because of this and make
the right choise wrt NNE vs. adding a catch block.

Reviewed By: @bertmaher

Differential Revision: D1400014