don't recompute files with errors
Summary: Files containing errors are systematically recomputed every time a file changes.
They are re-introduced in the pipeline in different places (depending on where they failed).
For example, if a file failed type-checking (as opposed to parsing, or naming), it is re-introduced
in the type-checking phase to see if the error went away.

The amount of hack errors present at any given time is relatively low in practice (Hack is a zero error system).
Because of that: recomputing all the files with an error was reasonably fast.

However, we are about to introduce a system (HH_FIXME) that let's the programmer ignore some errors.
The amount of files with errors at any given time is probably going to grow: recomputing all the files
with an error every time a file changes is probably not going to scale very well.
Whether this system is a good idea or not, in general, it would be cleaner for the type-checker not
to recompute files when nothing related to that file has changed.

In this diff, the logic of the type-checker is now the following:
1) recompute all the files that changed on disk: keep the errors introduced, let's call them X
2) remove all the errors (from the last run) coming from files that have been recomputed
3) Keep all the errors (from the last run) coming from files that have not been recomputed: call them Y
4) Show concat(X, Y) to the user

This is correct if the following invariant holds:
"The position of an error should be located in the file that introduced the error".
In other words, if type-checking the class 'A' introduces an error in a file that doesn't contain 'A' we
are in trouble: the error would persist forever until we recheck the file containain 'A' for unrelated reasons.

I think this invariant is reasonable and I am pretty sure it holds right now.

Reviewed By: @elgenie

Differential Revision: D1387289