Make FuncEmitter nicer
Summary: Get rid of getters and setters where possible, and make it clear which
properties are used in a POD-like way, and which represent state that's
somehow managed by FuncEmitter methods.

This is not completely satisfying, but it's a pretty good stopping point
on the way to a readable all-the-things-loader---i.e., I'm not going to
touch it anymore.

Reviewed By: @alexmalyshev

Differential Revision: D1390885