Strengthen type constraint in ldLocInner
`ldLocInner` calls `ldLoc` to get the local, but doesn't constrain the
type beyond it being boxed or not.

When doing an `ldLoc` in a pseudoMain, an `LdGbl` is emitted instead of
an `LdLoc`. The type of the global is inferred using the constraint,
which in this case is 'Countness'. The guard for 'Countness' doesn't
guard on Uninit, which is required for Isset.

This changes the constraint passed to `ldLoc` from 'Countness' to
'CountnessInit' which preserves the "initness" of the local.