Add class_elt field and fix require extends bug
Summary: Previously, the methods actually implemented by a trait were not distinguished
from the methods that a trait would necessarily have by virtue of a require
extends declaration. This caused a required method that lived in a superclass
to sometimes override an implementation in a subclass (!).

This diff puts an extra breadcrumb in class_elt so we can tell them apart,
and makes nonrequired (i.e. really, really implemented by the trait) methods
take precedence over merely required.

ran unit tests locally and they passed, and on contbuild are taking a long time, so

Reviewed By: @jwatzman

Differential Revision: D1371534