Fix reuse of dead persistent sockets
This fixes a bug where dead persistent sockets could be re-used
because the return value of the recv() syscall wasn't properly
checked. Previously, if recv returned -1 to signal an error
without setting errno to EAGAIN, the socket was erroneously
reused. This brings the persistent socket reuse check to
parity with Zend.

Also removed an outdated comment.