Add fuzzy results to search
Summary: Add fuzzy results to search results - in this case, fuzzy just means if the result is a common subsequence of the query. (i.e. 'hlo' matches 'hello', but 'hol' doesn't).
This means search only indexes methods (we don't fuzzy search those, there are too many, it just seems to clutter results and slow things down) and then aggregates the fuzzy search results.

It adds ~2s to startup time.

--------------------------
| Fuzzy Search Algorithm |
--------------------------

The algorithm is supposed to be fast to index, compromising with a slightly slower search time.
Here are some examples of what should match:
 - `search_term` -> `result`
 - 'heo' -> 'hello'
 - 'awtbl' -> 'Awaitable'
 - 'mthdy' -> 'MothersDay'
 - 'bsfcbktcs' -> 'BaseFacebookTestCase'
 - 'ExactMatch' -> 'ExactMatch'

1. Indexing
----
The index we build at a high level is simple: we build a map from letters to terms which contain those letters (in important places).  This means the term `Awaitable` lies under only 'a' and 'w', but 'someClass_here' lies under 's', 'o', 'c', and 'h'.
Note: This excludes the filter part, but each index also basically has 4 indexes, one for each filter type (Class, Constant, Typedef, Function)
Also note: We only index Classes, Constants, Typedefs, and Functions - regular search handles methods (there are just too many to index this way)

To index:
a.) Run workers over files, so no two workers can touch the same file
b.) Each worker will add to two shared memory objects:
    i.) a list of keys in the file
    ii.) a map from key to definition (where a definition is a name, location, and type)
 - The purpose of the key list object is so we can easily build a global list of all keys as well as check which keys are added/removed each time we index the file
 - The purpose of the map is so we can find the actual definition once we know the key is related to the search term.
c.) The master process then aggregates all the information from shared memory, checking the difference in each file compared to last indexing
d.) The master process the main index (map from letters to terms which contain the letters)
e.) It also builds a term lookup table, where we can lookup a term by it's key to find out which files it is in - this is used so we can find the actual definition object later, once we match

2. Searching
----
Searching is pretty simple: Get the first letter of the query, find the list of terms with that letter, and check every single term if it matches our search query.

To search:
a.) Cut the first letter off the query
b.) Lookup the list of terms which contain that letter (we indexed it earlier)
    i.) If there was a filter specified, only get the terms matching that type
c.) Compare our query against each term in that list
d.) If our query is a common subsequence of one of the terms, add it to the results list
e.) While we compare our query, we keep track of a score:
    i.) The more gaps there are between matching characters, the higher the score
    ii.) The more letters which are the incorrect case, the higher the score
    iii.) Direct matches have a score of 0.
    iv.) The later we find out that we've matched, the higher the score
    v.) Longer terms have higher scores
 - If your query is very spread out in the result, it's less likely to be what you're looking for.
 - The shorter the result, the closer the length is to what you searched for, so it's more likely.
 - The earlier we find the match, the more likely it is to be what you're looking for.
 - The more cases of the characters that match, the more likely it is to be what you're searching for
f.) Sort the results by the score
g.) Merge the results with regular search results (which indexes methods as well) - only if no filter was specified

Reviewed By: @joelbeales

Differential Revision: D1418856