Refactor running psp hooks so it's not part of hphp_context_exit
Summary: Rather than sending a bool into the function---I want to make it
easier to deal with PSP as a separate step so we can possibly change whether it
runs after a request ends with a fatal.  The final defaulted arg was also only
for the PSP code, and the shutdown arg was always passed true.  I am changing
the RPCRequestHandler not to call the post-psp debugger hook because I think it
was probably actually not intentionally doing so (since it doesn't run psp).

Also removed the ExecutionContext* argument: hphp_session_exit seems to be
willing to assume we're doing it to g_context, and I think we can afford the
extra load from TLS.

Reviewed By: @jano

Differential Revision: D1405002