Fix a JIT crash in unsupported NewElem dims
Summary: I got a sandbox crash in ExternalThreadEventWaitHandle---the strangest
part is that it seemed to be trying to abandon a wait handle from the minstr
translator at JIT time.  Apparently in an opt build if we get an MW that we
don't understand we load a function to call out of these tables with an invalid
index and somehow we called some random code (I'm not 100% sure I understand
how).  The inspecting the junk in the crash led me to trying an FPassM with an
MW.  Just PUNT instead of crashing.

Reviewed By: @swtaarrs

Differential Revision: D1400770