Autocomplete: Rank results
Summary: Adds ranking for autocomplete results based on whether they are they type we expect or not.
Right now, this is a boolean. Eventually, this should be "how many steps away the value I want is from what I want" (with some kind of limit)

We determine whether something is the expected type by looking for the deepest fun call that we're autocompleting inside. We look at the type of that AND it's return type, and check Typing_subtype.is_subtype to decide if it's a good argument or not.

In json output mode, we include which args are good. In normal mode, we'll sort those to the top.

To achieve this:
-Instead of converting types to strings when we find results, we store results with their ty.
-When we autocomplete something, we store the position of the autocomplete and the typing_env if it's available.
-On the way up from the recursion where we checked arguments, we check the position of the autocomplete if it's inside an argument of the function we're checking. If it is, we store what type we were looking for.
-When results are requested, we take our results, the stored env, and the stored type (if we have them), and rank.

Minor stuff I did to make this work:
-new ID autocomplete hook is moved from naming to typing so that I can get an environment.
-left hint and ID autocomplete where it used to be for now but eventually, ID autocomplete needs to move to typing so global funcs get ranked.
-on new ID() completions, we now provide the constructor args to the user. (This lets us filter now too because we have a ty for it) I do some sketchy stuff here. (For new ID only) If no constructor ty exists, I make a fake one that takes no args. If a constructor does exist, I change the return type to be the class it's constructing. this make my ranking work, and otherwise, when we do typing_print, it shows void as the return type, which seems more incorrect.
-Refactored a bit as well. Stopped storing results as a Smap (more trouble than it's worth with classes/funs sharing the same name) and switched it to a list. In case we end up hitting our hooks more than once, is_autocomplete checks if we've got results yet.

More work still to do here refactoring wise and getting globals funcs ranked (in another diff)

Reviewed By: @pikatchu

Differential Revision: D1406315