Factor out most of tryInline()
Summary: We want to try incorporating inlining in the hottrace selector, with or
without including it still for profiling translations.  We'll start by
just refactoring as much of the decision logic as possible.

Notable logic that we can't just shunt into a function are the inlining
depth checks and the requirement that FPush* not be interpreted.  The
former I'll clean up (right now we do the stack depth and inlining depth
checks differently, which is silly).  The latter we'll try to solve by
just breaking tracelets after an FPush whenever it's interpreted, and
documenting this limitation (and, ideally, fixing it, someday).

Reviewed By: bsimmers, ottoni

Differential Revision: D1405929