Stop generating LdStack<Gen> after GuardStk<T> with the simplifier off.
Summary: With the simplifier off, HhbcTranslator::pop() ends up producing
LdStack<Gen> for PopR, even though we may have guarded on a known type
on the stack slot. This causes us to later emit a CheckType<BoxedCell>,
which wreaks havoc with guard relaxation as it tries to refine Str and
BoxedCell and dies on an assert.

Reviewed By: @ottoni

Differential Revision: D1406686