Rewrite inlining logic
Summary: Rewrite the inlining decision-making mechanism so that it can be more
easily used with arbitrary tracelet selectors.

This also fixes a bunch of bugs (or at least misbehaviors) in inlining:

- Before, if we were going to inline a stack of calls four-deep, we
  would produce tracelets for the rest of the callstack at each callee,
  which is O(n^2).  Let's do this in O(n) instead.

- Before, when we produced the tracelet for the first callee to analyze
  its shape, we would do so with inlining turned on.  If we had a stack
  of calls five-deep in total (which exceeds the depth limit), the
  tracelet for the first callee would successfully inline four calls,
  but we would fail when trying to inline that back into the original
  caller.  Instead, we should just elect not to inline the fifth-nested
  call.

We also change "inlining cost" to measure the cost of a single inlined
callstack, rather than the total cost of all inlining sites in a single
function.

This is a lot cleaner, and also we get to have a Decider class---who
doesn't want that?

We also get rid of the hot-calling-cold tight cost limit for now, for
two reasons:

  1/ It seems like we should be -more- willing to inline into hotter
     functions.
  2/ We have to track the first-level callee, which is not hard but
     slightly dirty and I didn't do it.

Reviewed By: @swtaarrs

Differential Revision: D1409050